T222 Reorganize the rendered plot
